+++ packages/charts/chart-web-components/src/horizontal-bar-chart/horizontal-bar-chart.bench.ts
@@ -1,0 +1,12 @@
+import { FluentDesignSystem } from '@fluentui/web-components';
+import { definition } from './horizontal-bar-chart.definition.js';
+
+definition.define(FluentDesignSystem.registry);
+
+const itemRenderer = () => {
+  const horizontalbarchart = document.createElement('fluent-horizontal-bar-chart');
+  return horizontalbarchart;
+};
+
+export default itemRenderer;
+export { tests } from '../utils/benchmark-wrapper.js';