+++ packages/charts/react-charting/src/components/AreaChart/AreaChart.base.tsx
@@ -470,7 +470,9 @@ export class AreaChartBase extends React.Component<IAreaChartProps, IAreaChartSt
       });
     }
 
-    this._isMultiStackChart = renderPoints.length > 1;
+    this._isMultiStackChart = !!(this.props.legendProps?.selectedLegends
+      ? renderPoints?.length >= 1
+      : renderPoints?.length > 1);
     return {
       renderData: renderPoints,
       maxOfYVal,